+++ hostd/CHANGELOG.md
@@ -1,5 +1,11 @@
 # hostd
 
+## 0.3.0
+
+### Minor Changes
+
+- Added app and installer icons.
+
 ## 0.2.0
 
 ### Minor Changes